Add slidingWindow and Time
