+++ web/ts/utilities/paginator.ts
@@ -38,6 +38,10 @@ export class Paginator<T> {
         return this;
     }
 
+    find(callback: (obj: T, i: number) => boolean): T {
+        return this.list.find(callback);
+    }
+
     hasElements() {
         return this.list.length > 0;
     }
